+++ packages/slate-plugins/src/elements/list/index.ts
@@ -7,3 +7,4 @@ export * from './queries';
 export * from './renderElementList';
 export * from './transforms';
 export * from './types';
+export * from './withList';
