+++ packages/slate-plugins/src/elements/list/queries/index.ts
@@ -1,1 +1,2 @@
 export * from './isList';
+export * from './isSelectionInListItem';
