+++ packages/slate-plugins/src/elements/list/transforms/index.ts
@@ -1,2 +1,5 @@
+export * from './moveListItemDown';
+export * from './moveListItemUp';
 export * from './toggleList';
+export * from './insertListItem';
 export * from './unwrapList';
