+++ packages/slate-plugins/src/elements/list/transforms/moveListItemUp.ts
@@ -1,0 +1,71 @@
+import { Ancestor, Editor, Path, Transforms } from 'slate';
+import { setDefaults } from '../../../common/utils/setDefaults';
+import { DEFAULTS_LIST } from '../defaults';
+import { ListOptions } from '../types';
+
+/**
+ * Move a list item next to its parent.
+ * The parent should be a list item.
+ */
+export const moveListItemUp = (
+  editor: Editor,
+  listNode: Ancestor,
+  listPath: number[],
+  listItemPath: number[],
+  options?: ListOptions
+) => {
+  const { li } = setDefaults(options, DEFAULTS_LIST);
+
+  const [listParentNode, listParentPath] = Editor.parent(editor, listPath);
+  if (listParentNode.type !== li.type) return;
+
+  const newListItemPath = Path.next(listParentPath);
+
+  // Move item one level up
+  Transforms.moveNodes(editor, {
+    at: listItemPath,
+    to: newListItemPath,
+  });
+
+  /**
+   * Move the next siblings to a new list
+   */
+  const listItemIdx = listItemPath[listItemPath.length - 1];
+  const siblingPath = [...listItemPath];
+  const newListPath = newListItemPath.concat(1);
+  let siblingFound = false;
+  let newSiblingIdx = 0;
+  listNode.children.forEach((n, idx) => {
+    if (listItemIdx < idx) {
+      if (!siblingFound) {
+        siblingFound = true;
+
+        Transforms.insertNodes(
+          editor,
+          {
+            type: listNode.type,
+            children: [],
+          },
+          { at: newListPath }
+        );
+      }
+
+      siblingPath[siblingPath.length - 1] = listItemIdx;
+      const newSiblingsPath = newListPath.concat(newSiblingIdx);
+      newSiblingIdx++;
+      Transforms.moveNodes(editor, {
+        at: siblingPath,
+        to: newSiblingsPath,
+      });
+    }
+  });
+
+  // Remove sublist if it was the first list item
+  if (!listItemIdx) {
+    Transforms.removeNodes(editor, {
+      at: listPath,
+    });
+  }
+
+  return true;
+};
