+++ packages/slate-plugins/src/handlers/reset-block-type/types.ts
@@ -6,7 +6,7 @@ export interface ResetBlockTypePluginRule {
    */
   types: string[];
 
-  hotkey: string | string[];
+  hotkey?: string | string[];
 
   /**
    * Additional condition to the rule.
