+++ stories/examples/playground.stories.tsx
@@ -77,6 +77,7 @@ import {
   withImageUpload,
   withInlineVoid,
   withLink,
+  withList,
   withNormalizeTypes,
   withTable,
   withToggleType,
@@ -206,6 +207,7 @@ export const Plugins = () => {
     withHistory,
     withTable(options),
     withLink(),
+    withList(options),
     withDeserializeHTML({ plugins }),
     withImageUpload(),
     withToggleType({ defaultType: options.p.type }),
