+++ stories/handlers/exit-break.stories.tsx
@@ -19,6 +19,7 @@ import {
   SoftBreakPlugin,
   TablePlugin,
   ToolbarElement,
+  withList,
   withToggleType,
   withTrailingNode,
   withTransforms,
@@ -41,6 +42,7 @@ export default {
 const withPlugins = [
   withReact,
   withHistory,
+  withList(options),
   withToggleType({ defaultType: options.p.type }),
   withTransforms(),
   withTrailingNode({ type: options.p.type }),
